create shipment as bulk

+++ database/migrations/2021_10_12_223013_create_shipments_table.php
@@ -16,8 +16,8 @@ class CreateShipmentsTable extends Migration
         Schema::create('shipments', function (Blueprint $table) {
             $table->id();
             // $table->enum('carrier', ['ARAMEX','DHL']);
-            $table->integer('internal_awb')->unique();
-            $table->integer('external_awb')->nullable();
+            $table->bigInteger('internal_awb')->unique();
+            $table->bigInteger('external_awb')->nullable();
 
             $table->string('reference1', 128)->nullable();
             $table->string('reference2', 128)->nullable();
